Fix "pretty formatting" of pointer implementations
Rust supports "pretty debug formatting" using `{:#?}` formatting string which
is an alternative to default `{:?}`, it breaks struct to multi line and,
adds indentation, etc.

https://github.com/AlexanderSchuetz97/sync-ptr/commit/263a169f11f5d41fcd00612c7c2c7833e6c279db
introduced change in how pointers are formatted here, and that change was
undesirable, we preferred to keep formatting without changes.

There was a fix which addressed compatibility concern but it did not touch
"pretty formatting" case.

This change brings formatting back to parity with what it was pre-263a169f11f5d41fcd00612c7c2c7833e6c279db